# improvements for Ln