Improving the Ln implementation by scaling to range [0.1, 1) and using Math.log
for an initial estimate. Special casing for numbers close to 1 for which
scaling would require too much precision.

Also improving the precision calculation in Log10, and replacing the division
by `ln(10)` to a multiplication (by adding `1/ln(10)` constant).

Benchmarks below (about the cases that are slower: these are high-precision
cases where I believe we weren't using enough internal precision - we don't
have tests with high precisions).

```
Ln/P2/S-100/D2-24       631µs ± 1%   129µs ± 1%   -79.52%  (p=0.029 n=4+4)
Ln/P2/S-10/D2-24        322µs ± 1%   114µs ± 1%   -64.71%  (p=0.029 n=4+4)
Ln/P2/S-2/D2-24         315µs ± 1%   124µs ± 1%   -60.65%  (p=0.029 n=4+4)
Ln/P2/S2/D2-24          313µs ± 1%   111µs ± 1%   -64.44%  (p=0.029 n=4+4)
Ln/P2/S10/D2-24         372µs ± 3%   121µs ± 2%   -67.60%  (p=0.029 n=4+4)
Ln/P2/S100/D2-24        691µs ± 2%   128µs ± 2%   -81.53%  (p=0.029 n=4+4)
Ln/P10/S-100/D2-24      897µs ± 2%   284µs ± 1%   -68.29%  (p=0.029 n=4+4)
Ln/P10/S-100/D10-24     886µs ± 2%   597µs ± 1%   -32.69%  (p=0.029 n=4+4)
Ln/P10/S-10/D2-24       604µs ± 1%   284µs ± 2%   -53.00%  (p=0.029 n=4+4)
Ln/P10/S-10/D10-24      609µs ± 2%   561µs ± 1%    -7.85%  (p=0.029 n=4+4)
Ln/P10/S-2/D2-24        564µs ± 3%   313µs ± 0%   -44.58%  (p=0.029 n=4+4)
Ln/P10/S-2/D10-24       640µs ± 6%   571µs ± 1%   -10.79%  (p=0.029 n=4+4)
Ln/P10/S2/D2-24         538µs ± 1%   301µs ± 1%   -43.96%  (p=0.029 n=4+4)
Ln/P10/S2/D10-24        569µs ± 3%   587µs ± 2%      ~     (p=0.114 n=4+4)
Ln/P10/S10/D2-24        576µs ± 4%   300µs ± 3%   -47.94%  (p=0.029 n=4+4)
Ln/P10/S10/D10-24       604µs ± 4%   623µs ± 2%      ~     (p=0.200 n=4+4)
Ln/P10/S100/D2-24       918µs ± 2%   286µs ± 2%   -68.84%  (p=0.029 n=4+4)
Ln/P10/S100/D10-24      927µs ± 3%   589µs ± 2%   -36.41%  (p=0.029 n=4+4)
Ln/P100/S-100/D2-24    36.8ms ± 2%  19.0ms ± 1%   -48.18%  (p=0.029 n=4+4)
Ln/P100/S-100/D10-24   35.0ms ± 1%  20.2ms ± 3%   -42.19%  (p=0.029 n=4+4)
Ln/P100/S-100/D100-24  36.2ms ± 1%  78.0ms ± 1%  +115.44%  (p=0.029 n=4+4)
Ln/P100/S-10/D2-24     33.5ms ± 2%  17.2ms ± 2%   -48.65%  (p=0.029 n=4+4)
Ln/P100/S-10/D10-24    35.1ms ± 2%  21.1ms ± 1%   -39.96%  (p=0.029 n=4+4)
Ln/P100/S-10/D100-24   34.9ms ± 3%  71.4ms ± 1%  +104.93%  (p=0.029 n=4+4)
Ln/P100/S-2/D2-24      33.8ms ± 1%  19.3ms ± 7%   -42.98%  (p=0.029 n=4+4)
Ln/P100/S-2/D10-24     34.4ms ± 1%  22.1ms ± 3%   -35.66%  (p=0.029 n=4+4)
Ln/P100/S-2/D100-24    34.8ms ± 1%  74.0ms ± 0%  +112.48%  (p=0.029 n=4+4)
Ln/P100/S2/D2-24       30.8ms ± 2%  18.4ms ± 2%   -40.25%  (p=0.029 n=4+4)
Ln/P100/S2/D10-24      34.6ms ± 1%  22.2ms ± 2%   -35.91%  (p=0.029 n=4+4)
Ln/P100/S2/D100-24     35.7ms ± 1%  78.7ms ± 2%  +120.44%  (p=0.029 n=4+4)
Ln/P100/S10/D2-24      34.3ms ± 1%  18.7ms ± 1%   -45.57%  (p=0.029 n=4+4)
Ln/P100/S10/D10-24     32.2ms ± 1%  21.9ms ± 4%   -32.03%  (p=0.029 n=4+4)
Ln/P100/S10/D100-24    32.8ms ± 1%  82.5ms ± 2%  +151.47%  (p=0.029 n=4+4)
Ln/P100/S100/D2-24     34.7ms ± 1%  18.9ms ± 1%   -45.52%  (p=0.029 n=4+4)
Ln/P100/S100/D10-24    34.0ms ± 2%  22.3ms ± 2%   -34.31%  (p=0.029 n=4+4)
Ln/P100/S100/D100-24   34.8ms ± 1%  98.0ms ± 1%  +181.17%  (p=0.029 n=4+4)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/apd/12)
<!-- Reviewable:end -->