@mjibson 2017-02-06
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 2 files reviewed at latest revision, 5 unresolved discussions.

---

*[const.go, line 40 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH-Eedl2YGUYyBO24I:-KcH-EeeTlAyGa9VtCdT:bg6rh26) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L40)):*
> ```Go
> 
> 	decimalE = makeConstWithPrecision(strE)
> 	// ln(2)
> ```

`ln(10)`?

---

*[const.go, line 43 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH-JUaeY17ws2HsnuJ:-KcH-JUaeY17ws2HsnuK:b-590r3u) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L43)):*
> ```Go
> 	decimalLog10 = makeConstWithPrecision(strLog10)
> 	// 1/ln(10)
> 	decimalInvLog10 = makeConstWithPrecision(strInvLog10)
> ```

maybe we should just rename these strings to Ln: `strLn10`, etc. They were this way because I copied them from the cockroach code.

---

*[const.go, line 45 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH07khU28bsyq8TMzg:-KcH07kiX_cvWx2ZXFuH:b16e069) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L45)):*
> ```Go
> 	decimalInvLog10 = makeConstWithPrecision(strInvLog10)
> 	// ln(2)
> 	decimalLog2 = makeConstWithPrecision(strLog2)
> ```

Is this used anymore? We should remove it if not since makeConstWithPrecision does some work during init that we could avoid.

---

*[context.go, line 516 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcHKMq_eF-21FBtBKhN:-KcHKMqalm9Q-RZ0pHUs:btx67u9) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/context.go#L516)):*
> ```Go
> 		p += p
> 	}
> 	p += 2
> ```

Do we still need this?

---

*[context.go, line 573 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcHMaUAzf7UysQjDMMN:-KcHMaUBsFhbAWAYayty:b-jf8i9s) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/context.go#L573)):*
> ```Go
> 		if err != nil {
> 			// We know that z is in a reasonable range; no errors should happen during conversion.
> 			panic(err)
> ```

Although I agree with the analysis that this shouldn't be possible, since this function can already return an error, I think it's safer to do return instead of panic. Same for below.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/12#-:-KcHSExQZ98zhF-3XJEe:bnfp4nl)*
<!-- Sent from Reviewable.io -->