@mjibson 2017-02-06
Review status: 0 of 2 files reviewed at latest revision, 5 unresolved discussions.

---

*[context.go, line 516 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcHKMq_eF-21FBtBKhN:-KcJTi4zxPGn_mKYRxL4:b-a6qz3x) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/context.go#L516)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

I am reconsidering this; I will work on it a bit more.

I think we only need this double precision when `z` is close to 1, because that is when the scale of `e^a` (and `z`) and that of `a` (and `ln(z)`) differ by p, e.g. z=1.00001 vs ln(z)=~0.00001. 

But for these cases, the Taylor series for `ln(1+x)` should be much better as the terms are powers of `x` which decrease rapidly for small `x`.

I will experiment with this, and see where we have a good cut-off performance-wise.
</blockquote></details>

I wasn't talking about doubling the precision, just adding `2` to p. No need to do more work here for this commit.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/12)*
<!-- Sent from Reviewable.io -->