@RaduBerinde 2017-02-06
*[context.go, line 516 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcHKMq_eF-21FBtBKhN:-KcJW9hplCVNyYtNMLZk:bccdu8e) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/context.go#L516)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

I wasn't talking about doubling the precision, just adding `2` to p. No need to do more work here for this commit.
</blockquote></details>

I know, but the doubling of precision was bothering me anyway. The `+2` is there because the calculated precision applies to one iteration, but with multiple iterations the errors add up so we add a couple extra digits to compensate.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/12)*
<!-- Sent from Reviewable.io -->