@RaduBerinde 2017-02-06
Updated. Added a different method (power series) for values close to 1, and with this I was able to keep the precision down. The benchmarks look much better now.

All these tests are fantastic!  I tried a bunch of variants that turned out to have precision issues.


```
name                   old time/op  new time/op  delta
Ln/P2/S-100/D2-24       665µs ± 2%   104µs ± 1%  -84.30%  (p=0.029 n=4+4)
Ln/P2/S-10/D2-24        338µs ± 1%    79µs ± 1%  -76.55%  (p=0.029 n=4+4)
Ln/P2/S-2/D2-24         334µs ± 2%    92µs ± 2%  -72.32%  (p=0.029 n=4+4)
Ln/P2/S2/D2-24          338µs ± 5%    83µs ± 3%  -75.53%  (p=0.029 n=4+4)
Ln/P2/S10/D2-24         383µs ± 2%    86µs ± 1%  -77.59%  (p=0.029 n=4+4)
Ln/P2/S100/D2-24        709µs ± 2%   101µs ± 4%  -85.68%  (p=0.029 n=4+4)
Ln/P10/S-100/D2-24      962µs ±12%   234µs ± 0%  -75.73%  (p=0.029 n=4+4)
Ln/P10/S-100/D10-24     923µs ± 1%   264µs ± 0%  -71.41%  (p=0.029 n=4+4)
Ln/P10/S-10/D2-24       636µs ± 2%   240µs ± 1%  -62.28%  (p=0.029 n=4+4)
Ln/P10/S-10/D10-24      624µs ± 2%   241µs ± 2%  -61.39%  (p=0.029 n=4+4)
Ln/P10/S-2/D2-24        581µs ± 2%   266µs ± 3%  -54.16%  (p=0.029 n=4+4)
Ln/P10/S-2/D10-24       639µs ± 3%   231µs ± 2%  -63.88%  (p=0.029 n=4+4)
Ln/P10/S2/D2-24         556µs ± 3%   254µs ± 2%  -54.34%  (p=0.029 n=4+4)
Ln/P10/S2/D10-24        601µs ± 6%   233µs ± 1%  -61.19%  (p=0.029 n=4+4)
Ln/P10/S10/D2-24        614µs ± 8%   239µs ± 2%  -61.12%  (p=0.029 n=4+4)
Ln/P10/S10/D10-24       618µs ± 3%   272µs ± 0%  -56.03%  (p=0.029 n=4+4)
Ln/P10/S100/D2-24       952µs ± 1%   231µs ± 6%  -75.75%  (p=0.029 n=4+4)
Ln/P10/S100/D10-24      942µs ± 1%   254µs ± 2%  -73.05%  (p=0.029 n=4+4)
Ln/P100/S-100/D2-24    37.4ms ± 0%  18.8ms ± 6%  -49.61%  (p=0.029 n=4+4)
Ln/P100/S-100/D10-24   36.5ms ± 1%  16.2ms ± 5%  -55.55%  (p=0.029 n=4+4)
Ln/P100/S-100/D100-24  39.0ms ± 3%  17.1ms ± 1%  -56.04%  (p=0.029 n=4+4)
Ln/P100/S-10/D2-24     34.9ms ± 2%  15.6ms ± 1%  -55.46%  (p=0.029 n=4+4)
Ln/P100/S-10/D10-24    37.5ms ±10%  16.5ms ± 5%  -56.06%  (p=0.029 n=4+4)
Ln/P100/S-10/D100-24   34.8ms ± 1%  16.9ms ± 3%  -51.29%  (p=0.029 n=4+4)
Ln/P100/S-2/D2-24      35.3ms ± 4%  18.4ms ± 2%  -47.77%  (p=0.029 n=4+4)
Ln/P100/S-2/D10-24     35.3ms ± 1%  18.0ms ± 1%  -49.18%  (p=0.029 n=4+4)
Ln/P100/S-2/D100-24    37.5ms ± 2%  18.2ms ±15%  -51.46%  (p=0.029 n=4+4)
Ln/P100/S2/D2-24       32.4ms ± 1%  17.7ms ± 1%  -45.48%  (p=0.029 n=4+4)
Ln/P100/S2/D10-24      36.8ms ± 2%  20.1ms ± 8%  -45.42%  (p=0.029 n=4+4)
Ln/P100/S2/D100-24     37.6ms ± 4%  17.1ms ± 3%  -54.50%  (p=0.029 n=4+4)
Ln/P100/S10/D2-24      36.2ms ± 2%  18.8ms ± 2%  -48.02%  (p=0.029 n=4+4)
Ln/P100/S10/D10-24     33.8ms ± 1%  18.0ms ± 1%  -46.65%  (p=0.029 n=4+4)
Ln/P100/S10/D100-24    34.8ms ± 2%  16.1ms ± 1%  -53.75%  (p=0.029 n=4+4)
Ln/P100/S100/D2-24     36.8ms ± 1%  18.5ms ± 4%  -49.73%  (p=0.029 n=4+4)
Ln/P100/S100/D10-24    36.4ms ± 2%  18.7ms ± 1%  -48.65%  (p=0.029 n=4+4)
Ln/P100/S100/D100-24   37.5ms ± 7%  20.4ms ± 4%  -45.77%  (p=0.029 n=4+4)
```

---

Review status: 0 of 2 files reviewed at latest revision, 5 unresolved discussions.

---

*[const.go, line 40 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH-Eedl2YGUYyBO24I:-KcK0FWF5-nF5Jen4jMY:b-896fix) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L40)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

`ln(10)`?
</blockquote></details>

Done.

---

*[const.go, line 43 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH-JUaeY17ws2HsnuJ:-KcK0FkbEctE995sfX33:b-896fix) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L43)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

maybe we should just rename these strings to Ln: `strLn10`, etc. They were this way because I copied them from the cockroach code.
</blockquote></details>

Done.

---

*[const.go, line 45 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcH07khU28bsyq8TMzg:-KcK0G-5WKxflo4ORzL2:b-896fix) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/const.go#L45)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

Is this used anymore? We should remove it if not since makeConstWithPrecision does some work during init that we could avoid.
</blockquote></details>

Done.

---

*[context.go, line 573 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/12#-KcHMaUAzf7UysQjDMMN:-KcK0Pb3UEPAm9XDFwQL:b-896fix) ([raw file](https://github.com/cockroachdb/apd/blob/f3d997718f79320cd56e234742a0022ae17534d7/context.go#L573)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

Although I agree with the analysis that this shouldn't be possible, since this function can already return an error, I think it's safer to do return instead of panic. Same for below.
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/12#-:-KcK0kxcnOf7W081AlJ-:bx0dgge)*
<!-- Sent from Reviewable.io -->